Refactoriza la estructura como flutter recomienda
