+++ lib/api/services/noticia_service.dart
@@ -1,90 +1,141 @@
-import 'package:dcristaldo/data/noticia_repository.dart';
+import 'dart:async';
 import 'package:dcristaldo/domain/noticia.dart';
+import 'package:dio/dio.dart';
+import 'package:dcristaldo/constants.dart';
 
 class NoticiaService {
-  final NoticiaRepository _repository = NoticiaRepository();
-
-  /// Obtener todas las noticias sin paginación
-  Future<List<Noticia>> obtenerNoticias() async {
-    // Obtener todas las noticias desde el repositorio
-    final noticias = await _repository.obtenerNoticias();
-
-    // Validar que las noticias tengan datos válidos
-    for (final noticia in noticias) {
-      validarNoticia(noticia);
-    }
-
-    return noticias;
-  }
-
+  final Dio _dio = Dio(
+          BaseOptions(
+            connectTimeout: const Duration(seconds: CategoryConstants.timeoutSeconds),
+            receiveTimeout:const  Duration(seconds: CategoryConstants.timeoutSeconds),
+          ),
+        );
+  final String path=NewsConstants.noticiasEndpoint;
   /// Crear una nueva noticia
   Future<Noticia> crearNoticia(Noticia noticia) async {
-    try{
-      // Validar que los datos de la noticia sean válidos
-      validarNoticia(noticia);
-      // Enviar la noticia al repositorio
-      return await _repository.crearNoticia(noticia);
-    }catch(e){
+    try {
+      final response = await _dio.post(
+        path,
+        data: {
+          'titulo': noticia.titulo,
+          'descripcion': noticia.descripcion,
+          'fuente': noticia.fuente,
+          'publicadaEl': noticia.publicadaEl.toIso8601String(),
+          'urlImage': noticia.imagenUrl,
+          'categoriaId': noticia.categoriaId,
+        },
+      );
+      if (response.statusCode != 201) {
+        throw Exception('Error al crear la noticia: ${response.statusMessage}');
+      }else{
+        return Noticia.fromJson(response.data);
+      }
+    } catch (e) {
       throw Exception('Error al crear la noticia: $e');
     }
-    
+  }
+
+  /// Leer todas las noticias
+  Future<List<Noticia>> obtenerNoticias() async {
+    try {
+      final response = await _dio.get(path);
+
+      if (response.statusCode == 200) {
+        final List<dynamic> data = response.data ?? [];
+        return data.map((json) => Noticia.fromJson(json)).toList();
+      } else {
+        throw Exception('Error al obtener las noticias: ${response.statusMessage}');
+      }
+    } catch (e) {
+      throw Exception('Error al obtener las noticias: $e');
+    }
+  }
+
+  /// Leer una noticia por ID
+  Future<Noticia> obtenerNoticiaPorId(String id) async {
+    try {
+      final response = await _dio.get('$path/$id');
+
+      if (response.statusCode == 200) {
+        return Noticia.fromJson(response.data);
+      } else {
+        throw Exception('Error al obtener la noticia: ${response.statusMessage}');
+      }
+    } catch (e) {
+      throw Exception('Error al obtener la noticia: $e');
+    }
   }
 
   /// Actualizar una noticia
   Future<void> actualizarNoticia(String id, Noticia noticia) async {
-    validarNoticia(noticia);
-    await _repository.actualizarNoticia(id, noticia);
+    try {
+      final response = await _dio.put(
+        '$path/$id',
+        data: {
+          'titulo': noticia.titulo,
+          'descripcion': noticia.descripcion,
+          'fuente': noticia.fuente,
+          'publicadaEl': noticia.publicadaEl.toIso8601String(),
+          'urlImage': noticia.imagenUrl,
+          'categoriaId': noticia.categoriaId,
+        },
+      );
+
+      if (response.statusCode != 200) {
+        throw Exception('Error al actualizar la noticia: ${response.statusMessage}');
+      }
+    } catch (e) {
+      throw Exception('Error al actualizar la noticia: $e');
+    }
   }
 
   /// Eliminar una noticia
   Future<void> eliminarNoticia(String id) async {
-    // Eliminar la noticia desde el repositorio
-    await _repository.eliminarNoticia(id);
-  }
+    try {
+      final response = await _dio.delete('$path/$id');
 
-  /// Validar los datos de una noticia
-  void validarNoticia(Noticia noticia) {
-    if (noticia.publicadaEl.isAfter(DateTime.now())) {
-      throw ArgumentError('La fecha de publicación no puede estar en el futuro.');
-    }
-    if (noticia.titulo.trim().isEmpty) {
-      throw ArgumentError('El título de la noticia no puede estar vacío.');
-    }
-    if (noticia.descripcion.trim().isEmpty) {
-      throw ArgumentError('La descripción de la noticia no puede estar vacía.');
-    }
-    if (noticia.fuente.trim().isEmpty) {
-      throw ArgumentError('La fuente de la noticia no puede estar vacía.');
-    }
-    if (noticia.imagenUrl.trim().isEmpty) {
-      throw ArgumentError('La URL de la imagen no puede estar vacía.');
+      if (response.statusCode != 200 && response.statusCode != 204) {
+        throw Exception('Error al eliminar la noticia: ${response.statusMessage}');
+      }
+    } catch (e) {
+      throw Exception('Error al eliminar la noticia: $e');
     }
   }
-
-  /// Obtener noticias paginadas
-  /*Future<List<Noticia>> obtenerNoticiasPaginadas({
+  
+  /*Future<List<Noticia>> obtenerNoticiasDesdeApi({
     required int numeroPagina,
-    required int tamanoPaginaConst,
+    required int tamanoPagina,
   }) async {
-    // Validar que numeroPagina sea mayor o igual a 1 y tamañoPagina mayor a 0
-    if (numeroPagina < 1) {
-      throw ArgumentError('El número de página debe ser mayor o igual a 1.');
-    }
-    if (tamanoPaginaConst <= 0) {
-      throw ArgumentError('El tamaño de página debe ser mayor a 0.');
-    }
+    try {
+      // Realizar la solicitud HTTP a la API
+      final response = await _dio.get(
+        NewsConstants.url,
+      );
 
-    // Obtener las noticias paginadas desde el repositorio
-    final noticias = await _repository.obtenerNoticias(
-      numeroPagina: numeroPagina,
-      tamanoPagina: tamanoPaginaConst,
-    );
-
-    // Validar que las noticias tengan datos válidos
-    for (final noticia in noticias) {
-      validarNoticia(noticia);
+      if (response.statusCode == 200) {
+        final List<dynamic> data = response.data ?? [];
+        return data.map((json) => Noticia.fromJson(json)).toList();
+      } else {
+        throw Exception('Error al obtener las noticias: ${response.statusCode}');
+      }
+    } catch (e) {
+      throw Exception('Error al realizar la solicitud: $e');
     }
-
-    return noticias;
+  }
+  Future<void> crearNoticia(Noticia noticia) async {
+    try {
+      await _dio.post(
+        NewsConstants.url, // Cambia esta URL por la de tu API
+        data: {
+          'titulo': noticia.titulo,
+          'descripcion': noticia.descripcion,
+          'fuente': noticia.fuente,
+          'publicadaEl': noticia.publicadaEl.toIso8601String(),
+          'urlImage': noticia.imagenUrl,
+        },
+      );
+    } catch (e) {
+      throw Exception('Error al crear la noticia: $e');
+    }
   }*/
 }
