+++ lib/data/categoria_repository.dart
@@ -1,83 +1,63 @@
+import 'package:dcristaldo/api/services/categoria_service.dart';
 import 'package:dcristaldo/domain/categoria.dart';
-import 'package:dio/dio.dart';
-import 'package:dcristaldo/constants.dart';
 import 'package:dcristaldo/exceptions/api_exception.dart';
-
 class CategoriaRepository {
-  final Dio _dio = Dio();
-  final String path=CategoryConstants.categoriaEndpoint;
+  final CategoriaService _service= CategoriaService();
+  
   // Obtener todas las categorías
-   Future<List<Categoria>> obtenerCategorias() async {
+  Future<List<Categoria>> obtenerCategorias() async {
     try {
-      final response = await _dio.get(path);
-
-      if (response.statusCode == 200) {
-        final List<dynamic> categoriasJson = response.data;
-        return categoriasJson.map((json) => Categoria.fromJson(json)).toList();
-      } else {
-        throw ApiException(
-          'Error al obtener las categorías',
-          statusCode: response.statusCode,
-        );
-      }
-    } on DioException catch (e) {
-      throw ApiException(
-        'Error al conectar con la API de categorías: $e',
-        statusCode: e.response?.statusCode,
-      );
+      return await _service.obtenerCategorias();
     } catch (e) {
-      throw ApiException('Error desconocido: $e');
+      if (e is ApiException) {
+        // Propaga el mensaje contextual de ApiException
+        rethrow;
+      } else {
+        throw Exception('Error desconocido: $e');
+      }
     }
   }
 
   // Crear una nueva categoría
   Future<void> crearCategoria(Categoria categoria) async {
     try {
-      final response = await _dio.post(
-        path,
-        data: categoria.toJson(),
-      );
-      if (response.statusCode != 201) {
-        throw ApiException(
-          'Error al crear la categoría',
-          statusCode: response.statusCode,
-        );
-      }
+      await _service.crearCategoria(categoria);
     } catch (e) {
-      throw ApiException('Error al conectar con la API de categorías: $e');
+      if (e is ApiException) {
+        // Propaga el mensaje contextual de ApiException
+        throw Exception('Error en el servicio de categorías: ${e.message}');
+      } else {
+        throw Exception('Error desconocido: $e');
+      }
     }
   }
 
   // Actualizar una categoría existente
   Future<void> actualizarCategoria(String id, Categoria categoria) async {
     try {
-      final response = await _dio.put(
-        '$path/$id',
-        data: categoria.toJson(),
-      );
-      if (response.statusCode != 200) {
-        throw ApiException(
-          'Error al editar la categoría',
-          statusCode: response.statusCode,
-        );
-      }
+      await _service.actualizarCategoria(id, categoria);
     } catch (e) {
-      throw ApiException('Error al conectar con la API de categorías: $e');
+      if (e is ApiException) {
+        // Propaga el mensaje contextual de ApiException
+        throw Exception('Error en el servicio de categorías: ${e.message}');
+      } else {
+        throw Exception('Error desconocido: $e');
+      }
     }
   }
+
 
   // Eliminar una categoría
   Future<void> eliminarCategoria(String id) async {
     try {
-      final response = await _dio.delete('$path/$id');
-      if (response.statusCode != 200 && response.statusCode != 204) {
-        throw ApiException(
-          'Error al eliminar la categoría',
-          statusCode: response.statusCode,
-        );
-      }
+      await _service.eliminarCategoria(id);
     } catch (e) {
-      throw ApiException('Error al conectar con la API de categorías: $e');
+      if (e is ApiException) {
+        // Propaga el mensaje contextual de ApiException
+        rethrow;
+      } else {
+        throw Exception('Error desconocido: $e');
+      }
     }
   }
 }
